+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * `block::Atom::from_color()`, a const fn.
     * `block::BlockDef` can now be used to make block definitions that are looping animations.
       New items `block::Animation` and `block::BlockDef::new_animated()` support this functionality.
+    * `block::Resolution::from_log2()`, constructor from the logarithmic form.
     * `block::SetAttribute` is a modifier which overrides a single attribute of a block, instead of all of them.
       It replaces `Modifier::Attributes`.
     * `inv::InvInBlock` and `inv::IconRow` now have a complete set of getters for reading their data.